+++ _js/src/index.js
@@ -16,6 +16,7 @@
 
 import "@webcomponents/webcomponentsjs";
 import "intersection-observer";
+import ResizeObserver from "resize-observer-polyfill";
 import "web-animations-js";
 import smoothscroll from "smoothscroll-polyfill";
 import "../lib/request-idle-callback";
@@ -28,4 +29,5 @@ import "./drawer";
 import "./push-state";
 import "./katex";
 
+window.ResizeObserver = window.ResizeObserver || ResizeObserver;
 smoothscroll.polyfill();